Update CSS based on solution
It is not necessary to create more flex boxes when using blocks will suffice.

Eg: the title can just be a block.

Similarly, the text below each image did not have to be separated by a new nested flexbox.
Use max-width on the text+image to make sure the text does not extend beyond image width.
The text is automatically below the image because of default block layout.